# specs/openapi/admin-v0.4.2.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ScimCredentialCreated": {
        "properties": {
          "bearer_token": {
            "title": "Bearer Token",
            "type": "string"
          },
          "credential": {
            "$ref": "#/components/schemas/ScimCredentialOut"
          }
        },
        "required": [
          "credential",
          "bearer_token"
        ],
        "title": "ScimCredentialCreated",
        "type": "object"
      },
      "ScimCredentialOut": {
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "expires_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expires At"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "label": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Label"
          },
          "namespace": {
            "title": "Namespace",
            "type": "string"
          },
          "replaced_by_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Replaced By Id"
          },
          "revoked_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Revoked At"
          },
          "rotated_from_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rotated From Id"
          },
          "tenant_config_id": {
            "format": "uuid",
            "title": "Tenant Config Id",
            "type": "string"
          },
          "token_hint": {
            "title": "Token Hint",
            "type": "string"
          },
          "version": {
            "title": "Version",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "tenant_config_id",
          "namespace",
          "token_hint",
          "label",
          "rotated_from_id",
          "replaced_by_id",
          "version",
          "created_at",
          "expires_at",
          "revoked_at"
        ],
        "title": "ScimCredentialOut",
        "type": "object"
      },
      "ScimCredentialRotate": {
        "properties": {
          "expected_version": {
            "minimum": 1.0,
            "title": "Expected Version",
            "type": "integer"
          },
          "expires_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expires At"
          },
          "label": {
            "anyOf": [
              {
                "maxLength": 200,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Label"
          },
          "revoke_prior": {
            "default": true,
            "title": "Revoke Prior",
            "type": "boolean"
          }
        },
        "required": [
          "expected_version"
        ],
        "title": "ScimCredentialRotate",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    },
    "securitySchemes": {
      "APIKeyHeader": {
        "in": "header",
        "name": "X-Admin-Secret",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "description": "Provider-neutral decision receipts, bitemporal reconstruction, change-impact analysis, and auditable agent memory",
    "title": "Lians Decision Evidence API",
    "version": "0.4.2",
    "x-lians-api-surface": "admin"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/admin/enterprise/scim/tenants/{tenant_id}/credentials": {
      "get": {
        "operationId": "list_scim_credentials_v1_admin_enterprise_scim_tenants__tenant_id__credentials_get",
        "parameters": [
          {
            "in": "path",
            "name": "tenant_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Tenant Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "include_revoked",
            "required": false,
            "schema": {
              "default": false,
              "title": "Include Revoked",
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ScimCredentialOut"
                  },
                  "title": "Response List Scim Credentials V1 Admin Enterprise Scim Tenants  Tenant Id  Credentials Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "summary": "List Scim Credentials",
        "tags": [
          "admin",
          "enterprise",
          "scim"
        ]
      }
    },
    "/v1/admin/enterprise/scim/tenants/{tenant_id}/credentials/{credential_id}": {
      "delete": {
        "operationId": "revoke_scim_credential_v1_admin_enterprise_scim_tenants__tenant_id__credentials__credential_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "tenant_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Tenant Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "credential_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Credential Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "expected_version",
            "required": true,
            "schema": {
              "minimum": 1,
              "title": "Expected Version",
              "type": "integer"
            }
          },
          {
            "description": "Not accepted for this one-time or destructive operation; reconcile its authoritative resource before retrying.",
            "in": "header",
            "name": "Idempotency-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 255,
                  "minLength": 1,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Not accepted for this one-time or destructive operation; reconcile its authoritative resource before retrying.",
              "title": "Idempotency-Key"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "summary": "Revoke Scim Credential",
        "tags": [
          "admin",
          "enterprise",
          "scim"
        ]
      }
    },
    "/v1/admin/enterprise/scim/tenants/{tenant_id}/credentials/{credential_id}/rotate": {
      "post": {
        "operationId": "rotate_scim_credential_v1_admin_enterprise_scim_tenants__tenant_id__credentials__credential_id__rotate_post",
        "parameters": [
          {
            "in": "path",
            "name": "tenant_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Tenant Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "credential_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Credential Id",
              "type": "string"
            }
          },
          {
            "description": "Not accepted for this one-time or destructive operation; reconcile its authoritative resource before retrying.",
            "in": "header",
            "name": "Idempotency-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 255,
                  "minLength": 1,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Not accepted for this one-time or destructive operation; reconcile its authoritative resource before retrying.",
              "title": "Idempotency-Key"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ScimCredentialRotate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScimCredentialCreated"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "summary": "Rotate Scim Credential",
        "tags": [
          "admin",
          "enterprise",
          "scim"
        ]
      }
    },
...
  }
}